requirements.txt
Django==1.11.29
djangorestframework==3.9.4
pytz==2021.1
